// src/components/air-pricing/interface.ts
export interface AirPricing {
  airPricingID: string;
  aodid: string;
  aodName: string;
  aolid: string;
  aolName: string;
  commodityID: string;
  commodityName: string;
  currencyID: string;
  currencyAbbreviations: string;
  note: string;
  dateEffect: string;
  validityDate: string;
  freqDate: string;
  demAirPricing: string;
  detAirPricing: string;
  stoAirPricing: string;
  lclMinAirPricing: string;
  lclAirPricing: string;
  public: boolean;
  statusAirPricing: string;
  confirmDated: string;
  confirmByUser: string;
  dateInserted: string;
  insertedByUser: string;
  dateUpdated: string;
  updatedByUser: string;
  isDelete: boolean;
  dateDeleted: string;
  deleteByUser: string;
  airPricingDetailDTOs: { [key: string]: string };
  airPricingFeeDTOs: AirPricingFeeDTOs[];
}

export interface AirPricingDetailDTOs {
  [key: string]: string;
}
export interface AirPricingFeeDTOs {
  airPricingFeeID: string;
  feeID: string;
  feeName: string;
  feeNo: string;
  currencyID: string;
  currencyName: string;
  unitID: string;
  internationalCode: string;
  price: string;
}

export interface AirPricingTable extends Omit<AirPricing, 'airPricingID'> {
  key: string;
  searchAll: string;
}